update some changes

+++ src/main/java/com/filmrental/repository/AddressRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AddressRepository extends JpaRepository<Address, Short> {
+public interface AddressRepository extends JpaRepository<Address, Integer> {
     List<Address> findByCity_CityId(Integer cityId);
     List<Address> findByDistrict(String district);
     List<Address> findByPostalCode(String postalCode);
